Add a new setting to marginalia ":comments-indentation". This new setting is used to specify the size, in spaces, of the indentation used in the Clojure comments. This is required to properly format Markdown markup in comments. Let's take this example:
(defn foo
  "this is a test
    * first bullet
    * second bullet"
  [v])

Here, the comments indentation is two spaces. If this is not specified, the bullets won't be properly formatted in Markdown. This new option will strip the two spaces in front of the bullets to have the proper formatting in Markdown. This is particularly troublesome since Markdown uses spaces as special markup. Without this option, we would have to format comments that way, which is confusing to have the proper markup output:

(defn foo
  "this is a test
  * first bullet
  * second bullet"
  [v])